interpreter: move comparisons to utils folder (#2086)
We currently use comparisons, which are in the interpreter folder, in
the riscv_cf dialect, which makes for a circular import when I add
riscv_cf to `xdsl_run`. This PR moves that file, and centralises the
logic for evaluating the constant inputs.
