+++ xdsl/dialects/riscv_cf.py
@@ -11,7 +11,6 @@ from xdsl.dialects.riscv import (
     RISCVInstruction,
     RISCVRegisterType,
 )
-from xdsl.interpreters.comparisons import to_unsigned
 from xdsl.ir import Dialect, Operation, SSAValue
 from xdsl.irdl import (
     AttrSizedOperandSegments,
@@ -28,6 +27,7 @@ from xdsl.parser import Parser
 from xdsl.pattern_rewriter import RewritePattern
 from xdsl.printer import Printer
 from xdsl.traits import HasCanonicalisationPatternsTrait, IsTerminator
+from xdsl.utils.comparisons import to_signed, to_unsigned
 from xdsl.utils.exceptions import VerifyException
 
 
@@ -190,7 +190,7 @@ class ConditionalBranchOperation(IRDLOperation, RISCVInstruction, ABC):
         return op
 
     @abstractmethod
-    def const_evaluate(self, rs1: int, rs2: int) -> bool:
+    def const_evaluate(self, rs1: int, rs2: int, bitwidth: int) -> bool:
         pass
 
 
@@ -206,8 +206,10 @@ class BeqOp(ConditionalBranchOperation):
 
     name = "riscv_cf.beq"
 
-    def const_evaluate(self, rs1: int, rs2: int) -> bool:
-        return rs1 == rs2
+    def const_evaluate(self, rs1: int, rs2: int, bitwidth: int) -> bool:
+        lhs = to_unsigned(rs1, bitwidth)
+        rhs = to_unsigned(rs2, bitwidth)
+        return lhs == rhs
 
 
 @irdl_op_definition
@@ -222,8 +224,10 @@ class BneOp(ConditionalBranchOperation):
 
     name = "riscv_cf.bne"
 
-    def const_evaluate(self, rs1: int, rs2: int) -> bool:
-        return rs1 != rs2
+    def const_evaluate(self, rs1: int, rs2: int, bitwidth: int) -> bool:
+        lhs = to_unsigned(rs1, bitwidth)
+        rhs = to_unsigned(rs2, bitwidth)
+        return lhs != rhs
 
 
 @irdl_op_definition
@@ -238,8 +242,10 @@ class BltOp(ConditionalBranchOperation):
 
     name = "riscv_cf.blt"
 
-    def const_evaluate(self, rs1: int, rs2: int) -> bool:
-        return rs1 < rs2
+    def const_evaluate(self, rs1: int, rs2: int, bitwidth: int) -> bool:
+        lhs = to_signed(rs1, bitwidth)
+        rhs = to_signed(rs2, bitwidth)
+        return lhs < rhs
 
 
 @irdl_op_definition
@@ -254,8 +260,10 @@ class BgeOp(ConditionalBranchOperation):
 
     name = "riscv_cf.bge"
 
-    def const_evaluate(self, rs1: int, rs2: int) -> bool:
-        return rs1 >= rs2
+    def const_evaluate(self, rs1: int, rs2: int, bitwidth: int) -> bool:
+        lhs = to_signed(rs1, bitwidth)
+        rhs = to_signed(rs2, bitwidth)
+        return lhs >= rhs
 
 
 @irdl_op_definition
@@ -270,9 +278,10 @@ class BltuOp(ConditionalBranchOperation):
 
     name = "riscv_cf.bltu"
 
-    def const_evaluate(self, rs1: int, rs2: int) -> bool:
-        # TODO: use proper bitwidth here?
-        return to_unsigned(rs1, 32) < to_unsigned(rs2, 32)
+    def const_evaluate(self, rs1: int, rs2: int, bitwidth: int) -> bool:
+        lhs = to_unsigned(rs1, bitwidth)
+        rhs = to_unsigned(rs2, bitwidth)
+        return lhs < rhs
 
 
 @irdl_op_definition
@@ -287,9 +296,10 @@ class BgeuOp(ConditionalBranchOperation):
 
     name = "riscv_cf.bgeu"
 
-    def const_evaluate(self, rs1: int, rs2: int) -> bool:
-        # TODO: use proper bitwidth here?
-        return to_unsigned(rs1, 32) >= to_unsigned(rs2, 32)
+    def const_evaluate(self, rs1: int, rs2: int, bitwidth: int) -> bool:
+        lhs = to_unsigned(rs1, bitwidth)
+        rhs = to_unsigned(rs2, bitwidth)
+        return lhs >= rhs
 
 
 @irdl_op_definition
