+++ xdsl/interpreters/riscv.py
@@ -22,9 +22,9 @@ from xdsl.interpreter import (
     impl_cast,
     register_impls,
 )
-from xdsl.interpreters.comparisons import to_signed, to_unsigned
 from xdsl.ir import Attribute, SSAValue
 from xdsl.utils.bitwise_casts import convert_u32_to_f32
+from xdsl.utils.comparisons import to_signed, to_unsigned
 from xdsl.utils.exceptions import InterpretationError
 
 _T = TypeVar("_T")
