+++ xdsl/utils/comparisons.py
